Conflicts:
	components/bioformats
	components/blitz/resources/omero/api/IAdmin.ice
	components/blitz/resources/templates/factorymap.vm
	components/blitz/src/ome/services/blitz/impl/SharedResourcesI.java
	components/blitz/src/omero/grid/InteractiveProcessorI.java
	components/blitz/src/omero/grid/ParamsHelper.java
	components/blitz/src/omero/model/DetailsI.java
	components/blitz/test/ome/services/blitz/test/ManagedContextFixture.java
	components/insight/SRC/org/openmicroscopy/shoola/agents/measurement/view/MeasurementViewer.java
	components/insight/SRC/org/openmicroscopy/shoola/agents/measurement/view/MeasurementViewerComponent.java
	components/insight/SRC/org/openmicroscopy/shoola/agents/util/EditorUtil.java
	components/server/src/ome/security/basic/OmeroInterceptor.java
	components/tools/OmeroPy/src/omero/gateway/__init__.py
	components/tools/OmeroPy/test/integration/chgrp.py
	components/tools/OmeroPy/test/integration/ishare.py
	components/tools/OmeroPy/test/integration/library.py
	components/tools/OmeroWeb/omeroweb/webadmin/controller/experimenter.py
	components/tools/OmeroWeb/omeroweb/webadmin/controller/group.py
	components/tools/OmeroWeb/omeroweb/webadmin/templates/webadmin/experimenters.html
	components/tools/OmeroWeb/omeroweb/webadmin/views.py
	components/tools/OmeroWeb/omeroweb/webclient/templates/webclient/index/index_context.html
	components/tools/OmeroWeb/omeroweb/webclient/views.py
	components/tools/OmeroWeb/omeroweb/webclient/webclient_gateway.py
	components/tools/OmeroWeb/omeroweb/webgateway/views.py

+++ components/insight/SRC/org/openmicroscopy/shoola/env/data/OmeroMetadataServiceImpl.java
@@ -180,13 +180,11 @@ class OmeroMetadataServiceImpl
 		throws DSOutOfServiceException, DSAccessException 
 	{
 		if (annotation == null || object == null) return;
-		ExperimenterData exp = getUserDetails();
-		if (exp == null) return;
 		IObject ho = gateway.findIObject(ctx, annotation.asIObject());
-		if (ho == null) return;
+		if (ho == null || !gateway.canDelete(ho)) return;
 		IObject link = gateway.findAnnotationLink(ctx, object.getClass(),
-				       object.getId(), ho.getId().getValue(), exp.getId());
-		if (ho != null && link != null) {
+				       object.getId(), ho.getId().getValue(), -1);
+		if (link != null) {
 			try {
 				gateway.deleteObject(ctx, link);
 			} catch (Exception e) {
@@ -455,6 +453,7 @@ class OmeroMetadataServiceImpl
 	{			
 		String ioType = gateway.convertPojos(data).getName();
 		IObject ho = gateway.findIObject(ctx, ioType, data.getId());
+		if (ho == null) return;
 		ModelMapper.unloadCollections(ho);
 		IObject link = null;
 		boolean exist = false;
@@ -510,7 +509,6 @@ class OmeroMetadataServiceImpl
 		long id;
 		String ioType;
 		TagAnnotation ho;
-		IObject link = null;
 		if (ann instanceof TagAnnotationData && ann.isDirty()) {
 			TagAnnotationData tag = (TagAnnotationData) ann;
 			id = tag.getId();
